Add files via upload
This is the same version of the code that I sent via e-mail. It is able to solve the linearized and the non-linear equation. However, there are still issues with the inconsistent convergence plot for the case of the MP rule.  From now on, I will create branches and commit there first. Then I will request to pull into the main branch.